Improved midi parser. Handle some corner case.
In some case, a midi file can omit the status byte (the one that define the type
of a midi event). In such a case, the event is of the same kind of the previous
one. See http://www.somascape.org/midi/tech/spec.html#ctrlnums

The last_status_byte is initialised to an invalid value so to ensure an invalid
midi file won't be wrongly accepted.